include: configs: add get_fit_config for crocodile
Introduce a get_fit_config command that can be called by the boot
script (boot from eMMC) or CONFIG_BOOTCOMMAND (boot from QSPI).

This allows to abstract the fetch of the fit_config in the boot script,
that can be reused for multiple devices and, at the same time, to
bring the same logic in the QSPI image.

In the case of crocodile device, we select the fit_config based on the
board id, that is stored at the end of the QSPI header: 0x200.

Modified also qspiboot command to boot with the preselected fit_config.

Signed-off-by: Massimo Toscanelli <massimo.toscanelli@leica-geosystems.com>