beluga: stop autoboot via 'lll' key sequence
If for some reason, we receive a character on UART before autoboot is
started, we might end up in CLI and wait forever.

Limit keys to stop autoboot to 'lll' sequence in order to get a low
probability to accidentally stop autoboot.

This is similar to commit fdb9c2046ed3 ("crocodile: stop autoboot via
'lll' key sequence") for Crocodile.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>