arm64: dts: beluga: disable jr0 node as it is used by hab
JR0 is held in S-World since it might be used by HAB, therefore the node
has to be disabled.

Note, that this has to be done in the Kernel before it is synced back
into U-Boot, but without this exclusion SHA256 computations are broken
which prevents the FIT boot flow, see [1] for details.

There is a patch [2] for upstream Kernel which addresses the
reservation, but it will be included much later than we need since it
targets the linux-5.20.y kernel release.

Therefore, take this reservation earlier for Beluga machine, which
unblocks the FIT boot flow.

Link: [1]: https://lore.kernel.org/u-boot/AM6PR06MB46912207D9460CD9924F35DAA68B9@AM6PR06MB4691.eurprd06.prod.outlook.com/
Link: [2]: https://lore.kernel.org/all/20220608170223.1536594-1-festevam@denx.de/
Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>